tests/wpt/css-tests/css-transforms-1_dev/xhtml1print/transform_translate_zero_prefixed.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head>
    <title>CSS Transforms API Test: transform translate</title>
    <link href="http://www.intel.com" rel="author" title="Intel" />
    <link href="http://www.w3.org/TR/css-transforms-1/#transform-property" rel="help" />
    <link href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions" rel="help" />
    <meta content="dom" name="flags" />
    <meta content="Check if transform supports translate(0, 0)" name="assert" /> 
    <script src="/resources/testharness.js" type="text/javascript"></script>
    <script src="/resources/testharnessreport.js" type="text/javascript"></script>    
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script type="text/javascript">
        // Set the transform 
        document.getElementById("test").style.webkitTransform = "translate(0, 0)";
        /* document.getElementById("test").style.mozkitTransform = "translate(0, 0)";
        document.getElementById("test").style.msTransform = "translate(0, 0)";
        document.getElementById("test").style.oTransform = "translate(0, 0)";
        document.getElementById("test").style.transform = "translate(0, 0)"; */
 
        // Verify that the transform was set as expected
        var value = document.getElementById("test").style.getPropertyCSSValue("-webkit-transform").cssText;
        /* var value = document.getElementById("test").style.getPropertyCSSValue("-moz-transform").cssText;
        var value = document.getElementById("test").style.getPropertyCSSValue("-ms-transform").cssText;
        var value = document.getElementById("test").style.getPropertyCSSValue("-0-transform").cssText;
        var value = document.getElementById("test").style.getPropertyCSSValue("transform").cssText; */
        
        test(function() {assert_equals(value,
                        "translate(0px, 0px)",
                        "transform should be translate(0px, 0px)")}, 
                        "transform_translate_0_0");
    </script>    


</body></html>
